requirements.txt
rich==13.7.1
typer==0.12.3
inquirer==3.2.4
yaspin==3.0.2
moviepy==1.0.3
ffmpeg-python==0.2.0